pyproject.toml
[tool.poetry]
name = "cdsoffbot"
version = "0.1.0"
description = ""
authors = ["Mihir <mihirniyogi942@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.14"
gspread = "^5.7.0"
python-dotenv = "^0.21.0"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
